HW02 for  creating date-time filter

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -32,12 +32,16 @@ public class InMemoryMealRepository implements MealRepository {
             return meal;
         }
         // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        else if (repository.get(meal.getId()) != null && repository.get(meal.getId()).getUserId().equals(userId)) {
+            meal.setUserId(userId);
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        }
+        return null;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        if (repository.get(id).getUserId().equals(userId)) {
+        if (repository.get(id) != null && repository.get(id).getUserId().equals(userId)) {
             return repository.remove(id) != null;
         }
         return false;
@@ -45,7 +49,7 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public Meal get(int id, int userId) {
-        if (repository.get(id).getUserId().equals(userId)) {
+        if (repository.get(id) != null && repository.get(id).getUserId().equals(userId)) {
             return repository.get(id);
         }
         return null;
@@ -55,7 +59,7 @@ public class InMemoryMealRepository implements MealRepository {
     public List<Meal> getAll(int userId) {
         return repository.values().stream()
                 .filter(meal -> meal.getUserId().equals(userId))
-                .collect(Collectors.toCollection(ArrayList::new));
+                .collect(Collectors.toList());
     }
 }
